refactor(httpd): collapse dango-httpd server and schema bin into indexer-httpd
Move `dango-httpd`'s `run_server`, `config_app`, `not_found_handler`, and
the schema-export binary into `indexer-httpd`. `dango-httpd`'s source is
left as a placeholder lib.rs (deleted in commit 7) to keep the workspace
member valid for one more commit.

`indexer_httpd::server::run_server` now takes `shutdown_flag` and an
optional `port_sender`, replicating dango-httpd's signature. Its
`config_app` registers the full route set — index, up, sentry_raise,
blocks, graphql route over `FullQuery`/`FullSubscription`, optional
`/static` via `actix-files`, and a `not_found_handler` that serves
`{static_files_path}/404.html` when configured.

The schema-export binary moves from `dango/indexer/httpd/src/bin/` to
`indexer/httpd/src/bin/` and switches its imports to indexer-httpd. The
Justfile target is updated to `cargo run -p indexer-httpd --bin
build_graphql_schema`.

Callers (`dango/cli/src/start.rs`, `dango/mocks/httpd`, `dango/testing`)
flip their imports from `dango_httpd::*` to `indexer_httpd::*` and drop
the `dango-httpd` dependency. Two `tracing::error!` calls in moved
subscription resolvers gain `#[cfg(feature = "tracing")]` gates to match
indexer-httpd's optional-feature setup.

SDL byte-identical. `dango-httpd` is now an empty crate; commit 7
deletes it.

Part of phase 3 of the indexer refactor.
