--- dango/indexer/httpd/src/context.rs
@@ -1,1 +1,0 @@
-pub use indexer_httpd::context::FullContext as Context;
