+++ dango/indexer/httpd/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod context;
-pub mod graphql;
-pub mod server;
+// All content moved into `indexer-httpd` in phase 3 of testdata/refactor-indexer.md.
+// This crate is deleted in commit 7.
