+++ dango/mocks/httpd/Cargo.toml
@@ -13,7 +13,6 @@ version       = { workspace = true }
 anyhow                   = { workspace = true }
 dango-genesis            = { workspace = true }
 rand                     = { workspace = true }
-dango-httpd              = { workspace = true }
 dango-indexer-clickhouse = { workspace = true, features = ["tracing"] }
 dango-proposal-preparer  = { workspace = true }
 dango-testing            = { workspace = true }
